backend/requirements.txt
# requirements.txt - Python dependencies

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
firebase-admin==6.2.0

# Machine Learning
torch==2.1.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2

# HTTP client for LLM (Ollama)
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data validation
pydantic[email]==2.4.2

# Environment and configuration
python-dotenv==1.0.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# File processing
openpyxl==3.1.2
xlrd==2.0.1

# Logging and monitoring
python-json-logger==2.0.7

# Async utilities
asyncio-mqtt==0.13.0

# Testing (optional)
pytest==7.4.2
pytest-asyncio==0.21.1

# Development tools (optional)
black==23.9.1
flake8==6.1.0
mypy==1.6.1

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30
